+++ hooks/shortcode.php
@@ -1,6 +1,7 @@
 <?php
 
 use function Codevelopers\Markup\Helpers\Query\get_last_posts;
+use function Codevelopers\Markup\Helpers\TemplateTags\get_template_part;
 
 /**
  * Last posts Section shortcode.
@@ -18,5 +19,5 @@ add_shortcode('last-posts-section', function ($atts) {
     );
 
     $atts['posts'] = get_last_posts();
-    get_template_part('template-parts/last-posts-section', null, $atts);
+    get_template_part('last-posts-section', null, $atts);
 });