+++ index.php
@@ -1,5 +1,6 @@
 <?php
 
+use function Codevelopers\Markup\Helpers\TemplateTags\get_template_part;
 use function Codevelopers\Markup\Helpers\TemplateTags\site_main_row_container_class;
 use function Codevelopers\Markup\Helpers\TemplateTags\site_main_col_container_class;
 
@@ -11,16 +12,16 @@ get_header() ?>
             <main class="site-main" role="main">
                 <?php
                 if (have_posts()) :
-                    get_template_part('template-parts/content', 'grid');
-                    get_template_part('template-parts/paginator');
+                    get_template_part('content', 'grid');
+                    get_template_part('paginator');
                 else :
-                    get_template_part('template-parts/content', 'none');
+                    get_template_part('content', 'none');
                 endif;
                 ?>
             </main>
         </div>
 
-        <?php get_template_part('template-parts/sidebar', 'primary') ?>
+        <?php get_template_part('sidebar', 'primary') ?>
     </div>
 </div>