+++ template-full-width.php
@@ -4,13 +4,15 @@
  * Template Name: Full Width
  */
 
+use function Codevelopers\Markup\Helpers\TemplateTags\get_template_part;
+
 get_header() ?>
 
 <main class="site-main" role="main">
     <?php
     while (have_posts()) :
         the_post();
-        get_template_part('template-parts/content', 'page');
+        get_template_part('content', 'page');
     endwhile;
     ?>
 </main>